refactor(embed): split hybrid tests into sibling module
hybrid.rs sat at 298 lines, two below the 300-line cap. Move the
unit tests into a sibling hybrid_tests.rs (included via
#[cfg(test)] #[path] mod tests;) so the production module has room
for more fusion modes without further shaving (audit LOW
hybrid.rs:298).

No behaviour change.

Tracks: T6d8baabc-da68-4dd5-8781-b358ece69cbd
